generate.py: minor fixes

+++ generate.py
@@ -23,7 +23,7 @@ def clone_orekit_py_artifacts(version):
 
 orekit_version = [12, 0, 0, 0]  # for instance [11, 3, 3, 1] for Orekit version 11.3.3.1
 
-n_cpp_files_split = '99' # Number (in string representation) of C++ files to split. Alternatively 'separate' will create a C++ file for each Java class, can be a lot.
+n_cpp_files_split = '90' # Number (in string representation) of C++ files to split. Alternatively 'separate' will create a C++ file for each Java class, can be a lot.
 
 packages = ['java.io', 'java.util', 'java.text', 'org.orekit', 'org.orekit.rugged']
 vanilla_java_classes = [
@@ -68,9 +68,10 @@ classpath = []
 classes_exclude = ['org.hipparchus.util.RosenNumberPartitionIterator', 'org.orekit.compiler.plugin.DefaultDataContextPlugin']
 
 if __name__ == "__main__":
-    # First remove existing build folder to start from scratch
-    shutil.rmtree("build")
-    print("Deleted old ./build folder")
+    if os.path.isdir("build"):
+        # First remove existing build folder to start from scratch
+        shutil.rmtree("build")
+        print("Deleted old ./build folder")
 
     orekit_py_artifacts_tmp_dir = clone_orekit_py_artifacts(orekit_version)
     jar_folder = orekit_py_artifacts_tmp_dir.name
@@ -116,3 +117,5 @@ if __name__ == "__main__":
     cpp.jcc(args)
 
     orekit_py_artifacts_tmp_dir.cleanup()
+
+    print("To compile the Orekit C++ extension and install the Python package locally, run 'pip install .'")
